Move delete command to the domain layer
