+++ lib/pod/commands.rb
@@ -8,12 +8,13 @@ require_relative "commands/archive/runner"
 require_relative "commands/archive/output"
 require_relative "commands/dearchive/runner"
 require_relative "commands/dearchive/output"
+require_relative "commands/delete/runner"
+require_relative "commands/delete/output"
 
 require_relative "commands/init"
 require_relative "commands/podcasts"
 require_relative "commands/episodes"
 require_relative "commands/open"
-require_relative "commands/delete"
 require_relative "commands/sync"
 require_relative "commands/update"
