+++ lib/pod/commands/delete/output.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 module Pod
-  module Outputs
-    module Text
-      class Delete < ::Pod::Commands::BaseOutput
+  module Commands
+    module Delete
+      class Output < ::Pod::Commands::BaseOutput
         def call
           case @context[:details]
           when :podcast_deleted
