+++ spec/pod/commands/delete/output_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Pod::Outputs::Text::Delete do
+RSpec.describe Pod::Commands::Delete::Output do
   describe "#call" do
     context "when podcast was deleted" do
       it "generates the correct message" do
